install Tk with font rendering support in Conda environments on Linux
By default, Conda uses a version of Tk built without Linux font rendering support, which means that the OneTrainer interface will appear broken, with very tiny fonts and corrupt widget shapes. This problem does not affect Macs.

The upstream solution is to use Conda-Forge instead, which builds a better variant of Tk with full support for libXft, to render fonts beautifully on Linux.

We only apply this fix for Linux users, since the Xft package variant only exists for Linux.